+++ src/blocks/band-name-generator/render.php
@@ -1,0 +1,71 @@
+<?php
+/**
+ * Band Name Generator block render callback
+ */
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+$title = isset($attributes['title']) ? $attributes['title'] : 'Band Name Generator';
+$button_text = isset($attributes['buttonText']) ? $attributes['buttonText'] : 'Generate Band Name';
+
+$wrapper_attributes = get_block_wrapper_attributes([
+    'class' => 'extrachill-blocks-band-name-generator'
+]);
+
+$message_container = '<div class="extrachill-generator-message" style="display: none;"></div>';
+
+ob_start();
+?>
+<div <?php echo $wrapper_attributes; ?>>
+    <h3><?php echo esc_html($title); ?></h3>
+    <form class="extrachill-blocks-generator-form" data-generator-type="band">
+        <div class="form-group">
+            <label for="input"><?php _e('Your Name/Word:', 'extrachill-blog'); ?></label>
+            <input type="text" id="input" name="input" placeholder="<?php _e('Enter your name or word', 'extrachill-blog'); ?>" required>
+        </div>
+        <div class="form-group">
+            <label for="genre"><?php _e('Genre:', 'extrachill-blog'); ?></label>
+            <select id="genre" name="genre">
+                <option value="rock"><?php _e('Rock', 'extrachill-blog'); ?></option>
+                <option value="country"><?php _e('Country', 'extrachill-blog'); ?></option>
+                <option value="metal"><?php _e('Metal', 'extrachill-blog'); ?></option>
+                <option value="indie"><?php _e('Indie', 'extrachill-blog'); ?></option>
+                <option value="punk"><?php _e('Punk', 'extrachill-blog'); ?></option>
+                <option value="jam"><?php _e('Jam', 'extrachill-blog'); ?></option>
+                <option value="electronic"><?php _e('Electronic', 'extrachill-blog'); ?></option>
+                <option value="random"><?php _e('Random', 'extrachill-blog'); ?></option>
+            </select>
+        </div>
+        <div class="form-group">
+            <label for="number_of_words"><?php _e('Number of Words:', 'extrachill-blog'); ?></label>
+            <select id="number_of_words" name="number_of_words">
+                <option value="2"><?php _e('2 Words', 'extrachill-blog'); ?></option>
+                <option value="3"><?php _e('3 Words', 'extrachill-blog'); ?></option>
+                <option value="4"><?php _e('4 Words', 'extrachill-blog'); ?></option>
+            </select>
+        </div>
+        <div class="form-group">
+            <label>
+                <input type="checkbox" name="first-the" value="true">
+                <?php _e('Add "The" at the beginning', 'extrachill-blog'); ?>
+            </label>
+        </div>
+        <div class="form-group">
+            <label>
+                <input type="checkbox" name="and-the" value="true">
+                <?php _e('Add "& The" in the middle', 'extrachill-blog'); ?>
+            </label>
+        </div>
+        <button type="submit" class="button-1 button-medium"><?php echo esc_html($button_text); ?></button>
+    </form>
+    <?php echo $message_container; ?>
+    <div class="extrachill-blocks-generator-result" style="display: none;">
+        <div class="generated-name-wrap">
+            <em><?php _e('Your band name will appear here', 'extrachill-blog'); ?></em>
+        </div>
+    </div>
+</div>
+<?php
+return ob_get_clean();